requirements.txt
# =============================================================================
# AutoPaperGen - 学术论文初稿自动生成系统
# 依赖安装: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# 核心框架 - LLM 与 RAG
# -----------------------------------------------------------------------------

# LangChain - LLM 应用开发框架
langchain>=0.2.0,<0.4.0

# LangChain OpenAI 集成 - 用于 Qwen API 兼容调用
langchain-openai>=0.1.0,<0.3.0

# LangGraph - 工作流编排引擎，用于多 Agent 协作
langgraph>=0.1.0,<0.3.0

# LlamaIndex - RAG 检索增强生成框架
llama-index>=0.10.0,<0.13.0
llama-index-core>=0.10.0,<0.13.0

# -----------------------------------------------------------------------------
# 向量存储与检索
# -----------------------------------------------------------------------------

# ChromaDB - 向量数据库，用于文献和代码索引
chromadb>=0.4.0,<0.6.0

# LlamaIndex ChromaDB 集成
llama-index-vector-stores-chroma>=0.1.0,<0.5.0

# -----------------------------------------------------------------------------
# Qwen API 相关 - 阿里云通义千问
# -----------------------------------------------------------------------------

# DashScope - 阿里云 Qwen API SDK
dashscope>=1.14.0,<2.0.0

# OpenAI - 用于兼容 OpenAI API 格式调用 Qwen
openai>=1.0.0,<2.0.0

# -----------------------------------------------------------------------------
# 嵌入模型
# -----------------------------------------------------------------------------

# LlamaIndex HuggingFace 嵌入集成 - 本地嵌入模型支持
llama-index-embeddings-huggingface>=0.1.0,<0.6.0

# LlamaIndex OpenAI 嵌入集成 - 用于 Qwen 嵌入 API
llama-index-embeddings-openai>=0.1.0,<0.4.0

# -----------------------------------------------------------------------------
# 文件读取
# -----------------------------------------------------------------------------

# LlamaIndex 文件读取器 - 支持 PDF 等格式
llama-index-readers-file>=0.1.0,<0.5.0

# -----------------------------------------------------------------------------
# 模板渲染
# -----------------------------------------------------------------------------

# Jinja2 - 模板引擎，用于论文模板渲染
jinja2>=3.0.0,<4.0.0

# -----------------------------------------------------------------------------
# 配置与工具
# -----------------------------------------------------------------------------

# Pydantic - 数据验证与设置管理
pydantic>=2.0.0,<3.0.0

# Python Dotenv - 环境变量管理
python-dotenv>=1.0.0,<2.0.0

# PyYAML - YAML 配置文件解析
pyyaml>=6.0,<7.0

# Requests - HTTP 请求库
requests>=2.28.0,<3.0.0

# -----------------------------------------------------------------------------
# 代码解析（可选）- AST 感知的代码切分
# -----------------------------------------------------------------------------

# Tree-sitter 语言包 - 支持多种编程语言的 AST 解析
# 如遇到网络问题无法下载 parser，系统会自动降级使用文本切分
tree-sitter-language-pack>=1.0.0,<2.0.0

# -----------------------------------------------------------------------------
# 开发依赖（可选）
# -----------------------------------------------------------------------------

# pytest - 测试框架
# pytest>=7.0.0,<9.0.0
